AX: AXNotifications enum and AXComputedObjectAttributeCache can be extracted from AXObjectCache https://bugs.webkit.org/show_bug.cgi?id=297469 rdar://158425759
Reviewed by Joshua Hoffman.

AXObjectCache.h is included from a lot of places, we should remove definitions that aren't needed everywhere

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/accessibility/AXComputedObjectAttributeCache.cpp: Copied from Source/WebCore/accessibility/AXObjectCacheInlines.h.
(WebCore::AXComputedObjectAttributeCache::getIgnored const):
(WebCore::AXComputedObjectAttributeCache::setIgnored):
* Source/WebCore/accessibility/AXComputedObjectAttributeCache.h: Added.
* Source/WebCore/accessibility/AXLogger.cpp:
* Source/WebCore/accessibility/AXNotifications.h: Added.
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXComputedObjectAttributeCache::getIgnored const): Deleted.
(WebCore::AXComputedObjectAttributeCache::setIgnored): Deleted.
* Source/WebCore/accessibility/AXObjectCache.h:
(WebCore::CharacterOffset::isEqual const): Deleted.
(WebCore::CharacterOffset::debugDescription): Deleted.
(WebCore::AXComputedObjectAttributeCache::CachedAXObjectAttributes::CachedAXObjectAttributes): Deleted.
* Source/WebCore/accessibility/AXObjectCacheInlines.h:
(WebCore::CharacterOffset::debugDescription):
(WebCore::CharacterOffset::isEqual const):
* Source/WebCore/accessibility/AXSearchManager.cpp:
* Source/WebCore/accessibility/AccessibilityMenuList.cpp:
* Source/WebCore/accessibility/AccessibilityMenuListPopup.cpp:
* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
* Source/WebCore/accessibility/AccessibilityObject.cpp:
* Source/WebCore/accessibility/AccessibilityRenderObject.cpp:
* Source/WebCore/accessibility/ios/AXObjectCacheIOS.mm:
* Source/WebCore/accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
* Source/WebCore/accessibility/mac/AXObjectCacheMac.mm:
* Source/WebCore/page/ChromeClient.h:

Canonical link: https://commits.webkit.org/298796@main
