+++ Source/WebCore/Headers.cmake
@@ -831,10 +831,12 @@ set(WebCore_PRIVATE_FRAMEWORK_HEADERS
     Scripts/generate-log-declarations.py
 
     accessibility/AXAttributeCacheScope.h
+    accessibility/AXComputedObjectAttributeCache.h
     accessibility/AXCoreObject.h
     accessibility/AXGeometryManager.h
     accessibility/AXLogger.h
     accessibility/AXLoggerBase.h
+    accessibility/AXNotifications.h
     accessibility/AXObjectCache.h
     accessibility/AXObjectCacheInlines.h
     accessibility/AXSearchManager.h
