+++ Source/WebCore/accessibility/AccessibilityRenderObject.cpp
@@ -31,6 +31,7 @@
 
 #include "AXLogger.h"
 #include "AXLoggerBase.h"
+#include "AXNotifications.h"
 #include "AXObjectCache.h"
 #include "AXUtilities.h"
 #include "AccessibilityImageMapLink.h"
