+++ Source/WebCore/accessibility/ios/WebAccessibilityObjectWrapperIOS.mm
@@ -30,6 +30,7 @@
 
 #import "AXAttributeCacheScope.h"
 #import "AXLogger.h"
+#import "AXNotifications.h"
 #import "AXObjectCache.h"
 #import "AXObjectCacheInlines.h"
 #import "AXSearchManager.h"
