+++ Source/WebCore/accessibility/playstation/AXObjectCachePlayStation.cpp
@@ -26,6 +26,7 @@
 #include "config.h"
 #include "AXObjectCache.h"
 
+#include "AXNotifications.h"
 #include "AccessibilityObject.h"
 #include "Chrome.h"
 #include "ChromeClient.h"
